chore: fix mutex await across lint
